add data

+++ db/migrate/20220324163622_create_words.rb
@@ -3,7 +3,7 @@ class CreateWords < ActiveRecord::Migration[7.0]
     create_table :words do |t|
       t.string :word
       t.integer :anynum
-      t.integer :codeword
+      t.bigint :codeword
       t.string :a
       t.string :b
       t.string :c
